+++ service/controller/rbac/resource/externalresources/create.go
@@ -15,9 +15,6 @@ import (
 	"github.com/giantswarm/rbac-operator/service/controller/rbac/key"
 )
 
-// Ensures that Subjects with any sort of access in the organization namespace also have read access to
-// - releases (non-namespaced)
-// - app catalogs and app catalog entries in the default namespace
 func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	var err error
 
@@ -34,6 +31,79 @@ func (r *Resource) EnsureCreated(ctx context.Context, obj interface{}) error {
 	} else if len(orgRoleBindings.Items) == 0 {
 		return nil
 	}
+
+	// Ensure RoleBinding for default app catalogs access and ClusterRoleBinding for releases access
+	err = r.ensureAll(ctx, orgNamespace, orgRoleBindings)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	// Ensure RoleBindings for read/write access to cluster namespace resources
+	err = r.ensureClusterNamespaceAccess(ctx, orgNamespace, orgRoleBindings)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	return nil
+}
+
+// Ensures that Subjects with full read access to the org-namespace also have
+// - read-access to resources in the org cluster namespaces
+// Ensures that Subjects with admin access to the org-namespace also have
+// - write-access toresources in the org cluster namespaces
+func (r *Resource) ensureClusterNamespaceAccess(ctx context.Context, orgNamespace string, orgRoleBindings *rbacv1.RoleBindingList) error {
+	var err error
+
+	subjects := getUniqueSubjectsWithClusterRoleRef(orgRoleBindings, pkgkey.DefaultReadAllPermissionsName)
+	err = r.ensureRoleBindingToClusterRole(ctx, subjects, pkgkey.ReadClusterNamespaceAppsRole, orgNamespace, pkgkey.ReadClusterNamespaceAppsRoleBinding)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	subjects = getUniqueSubjectsWithClusterRoleRef(orgRoleBindings, pkgkey.ClusterAdminClusterRoleName)
+	err = r.ensureRoleBindingToClusterRole(ctx, subjects, pkgkey.WriteClusterNamespaceAppsRole, orgNamespace, pkgkey.WriteClusterNamespaceAppsRoleBinding)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	return nil
+}
+
+func (r *Resource) ensureRoleBindingToClusterRole(ctx context.Context, subjects []rbacv1.Subject, clusterRole string, namespace string, name string) error {
+	var err error
+
+	roleBinding := &rbacv1.RoleBinding{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "RoleBinding",
+			APIVersion: "rbac.authorization.k8s.io/v1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name: name,
+			Labels: map[string]string{
+				label.ManagedBy: project.Name(),
+			},
+			Namespace: namespace,
+		},
+		Subjects: subjects,
+		RoleRef: rbacv1.RoleRef{
+			APIGroup: "rbac.authorization.k8s.io",
+			Kind:     "ClusterRole",
+			Name:     clusterRole,
+		},
+	}
+
+	if err = r.createOrUpdateRoleBinding(ctx, roleBinding.Namespace, roleBinding); err != nil {
+		return microerror.Mask(err)
+	}
+
+	return nil
+}
+
+// Ensures that Subjects with any sort of access in the organization namespace also have read access to
+// - releases (non-namespaced)
+// - app catalogs and app catalog entries in the default namespace
+func (r *Resource) ensureAll(ctx context.Context, orgNamespace string, orgRoleBindings *rbacv1.RoleBindingList) error {
+	var err error
 
 	// Collect the subjects that need access
 	subjects := getUniqueSubjects(orgRoleBindings)